Fixed CCE in compareTo.
git-svn-id: https://svn.neo4j.org/components/neo/trunk@960 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ advanced/src/java/org/neo4j/impl/core/NodeProxy.java
@@ -133,7 +133,7 @@ class NodeProxy implements Node
 	public int compareTo( Object node )
 	{
 		Node n = (Node) node;
-		int ourId = (int) this.getId(), theirId = (int) n.getId();
+		long ourId = (long) this.getId(), theirId = (long) n.getId();
 		
 		if ( ourId < theirId )
 		{
